requirements.txt
accelerate==0.30.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
azure-ai-documentintelligence==1.0.0
azure-core==1.30.2
babel==2.17.0
beautifulsoup4==4.13.5
bitsandbytes==0.43.1
Brotli==1.1.0
bs4==0.0.2
catboost==1.2.8
certifi==2025.8.3
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
contourpy==1.3.3
courlan==1.3.2
cssselect==1.3.0
cycler==0.12.1
datasets==2.20.0
dateparser==1.2.2
dill==0.3.8
distro==1.9.0
docstring_parser==0.17.0
dotenv==0.9.9
einops==0.8.0
eval_type_backport==0.2.2
fastapi==0.115.0
feedfinder2==0.0.4
feedparser==6.0.11
ffmpy==0.6.1
filelock==3.13.1
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2024.5.0
gradio==5.46.1
gradio_client==1.13.1
graphviz==0.21
greenlet==3.2.4
groovy==0.1.2
h11==0.16.0
hf-xet==1.1.9
hf_transfer==0.1.9
htmldate==1.9.3
httpcore==1.0.9
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.34.4
idna==3.10
isodate==0.7.2
jieba3k==0.35.1
Jinja2==3.1.4
jiter==0.10.0
jusText==3.0.2
kiwisolver==1.4.9
lxml==5.4.0
lxml_html_clean==0.4.2
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.6
mdurl==0.1.2
mpmath==1.3.0
multidict==6.6.4
multiprocess==0.70.16
narwhals==2.3.0
networkx==3.3
newspaper3k==0.2.8
ninja==1.13.0
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
openai==1.106.1
orjson==3.11.3
pandas==2.3.2
peft==0.10.0
pillow==11.0.0
propcache==0.3.2
protobuf==6.32.0
pyarrow==21.0.0
pyarrow-hotfix==0.7
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydub==0.25.1
pyodbc==5.2.0
pyparsing==3.2.3
python-dotenv==1.1.1
python-multipart==0.0.9
pytz==2025.2
PyYAML==6.0.2
readability-lxml==0.8.4.1
regex==2025.9.1
requests==2.32.5
requests-file==2.1.0
rich==14.1.0
ruff==0.12.12
safehttpx==0.1.6
safetensors==0.4.3
scikit-learn==1.5.2
scipy==1.16.1
semantic-version==2.10.0
sentencepiece==0.2.1
sgmllib3k==1.0.0
shellingham==1.5.4
shtab==1.7.2
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.43
sqlglot==27.16.3
starlette==0.38.6
sympy==1.13.1
tabpfn==2.1.3
tinysegmenter==0.3
tld==0.13.1
tldextract==5.3.0
tokenizers==0.19.1
tomlkit==0.13.3
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
trafilatura==2.0.0
transformers==4.41.2
triton==3.1.0
trl==0.9.6
typeguard==4.4.4
typer==0.17.3
typing-inspection==0.4.1
tyro==0.9.31
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.30.6
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.1
pandas
tiktoken
mdformat
py-markdown-table
paramiko
azure-storage-blob
google-generativeai>=0.8.0
azure-ai-documentintelligence>=1.0.0
openpyxl>=3.1.0
xlrd==1.2.0
google-auth>=2.24.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.108.0
msal>=1.25.0
pymssql>=2.2.0
